Add SeparatorWriter interface
