+++ ltsview.go
@@ -6,13 +6,12 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/fatih/color"
 	"github.com/ymotongpoo/goltsv"
 )
 
 type LTSView struct {
 	Reader io.Reader
-	Writer io.Writer
+	Writer Writer
 
 	Keys  map[string]struct{}
 	Ikeys map[string]struct{}
@@ -51,30 +50,12 @@ func SortKeys(record map[string]string) []string {
 	return keys
 }
 
-func PrintLine(w io.Writer, k, v string) {
+func PrintLine(w Writer, k, v string) {
 	fmt.Fprintln(w, k+":", v)
 }
 
-func PrintSep(w io.Writer) {
-	fmt.Fprintln(w, "---")
-}
-
-type ColorableWriter struct {
-	io.Writer
-}
-
-func (w *ColorableWriter) Write(p []byte) (int, error) {
-	f := color.New(color.FgMagenta).SprintFunc()
-	for i := range p {
-		str := string(p[i])
-		if p[i] == ':' {
-			fmt.Fprint(w.Writer, str)
-			f = color.New(color.FgGreen).SprintFunc()
-			continue
-		}
-		fmt.Fprint(w.Writer, f(str))
-	}
-	return len(p), nil
+func PrintSep(w SeparatorWriter) {
+	w.WriteSeparator("---")
 }
 
 func ParseKeysByFlag(flag string) map[string]struct{} {
