+++ ltsview/main.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"flag"
-	"io"
 	"os"
 
 	"github.com/nabeken/go-ltsview"
@@ -10,13 +9,13 @@ import (
 
 var flagKeys = flag.String("k", "", "Specify comma-separated keys to show")
 var flagIkeys = flag.String("i", "", "Specify comma-separated keys to ignore")
-var flagNocolor = flag.Bool("n", false, "Set to true to disable colorize outputs")
+var flagNocolor = flag.Bool("n", false, "Set to true to disable colorized outputs")
 
 func main() {
 	flag.Parse()
-	var w io.Writer
+	var w ltsview.Writer
 	if *flagNocolor {
-		w = os.Stdout
+		w = ltsview.NewWriter(os.Stdout)
 	} else {
 		w = &ltsview.ColorableWriter{os.Stdout}
 	}
